running localhost simulation
