+++ simul/build_test.go
@@ -1,4 +1,4 @@
-package main
+package simul
 
 import (
 	"strconv"
